requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
matplotlib==3.8.3
seaborn==0.13.2
shap==0.44.1
imbalanced-learn==0.12.0
Pillow==10.2.0